refactor: Reorganize products, employees, inventory, and clients modules
- Improved separation of concerns in API client interactions.
