+++ src/features/portal/api/inventory/inventoryService.ts
@@ -1,4 +1,4 @@
-import api from '../../../utils/axiosConfig';
+import api from "../../../../utils/axiosConfig";
 
 export type MovementType = 'ENTRY' | 'EXIT';
