+++ src/features/portal/api/product/index.ts
@@ -1,0 +1,4 @@
+export * from './types';
+export { productService } from './productService';
+export { priceService } from './priceService';
+export { productReportService } from './reportService';
