+++ src/features/portal/api/sale/discountService.ts
@@ -1,4 +1,4 @@
-import api from '../../../utils/axiosConfig';
+import api from '../../../../utils/axiosConfig';
 
 export type DiscountType = 'PERCENTAGE' | 'FIXED_AMOUNT';
